testing/web-platform/tests/atomics/window-and-any-worker/wake/wake-zero/wake-zero-popup.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

// Copyright (C) 2017 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wake wakes zero waiters if that's what the count is.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    let popup = window.open("../empty.html", "empty");
    let test = this;
    popup.onmessage = function(e) {
      let worker = new Worker("wake-zero.js");
      worker.onmessage = test.step_func_done(function(e) {
        let ret = e.data[0];
        assert_equals(ret, "timed-out");
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia]);
      setTimeout(() => assert_equals(0, Atomics.wake(ia, 0, 0)), 500); // Wake zero
    }
    popup.postMessage([], "*");
  }, "atomics-dedicated-worker-wake-zero-from-popup");
});
</script>
